bug fixes and waiting for file lock release feature

+++ FolderMonitoringService/Services/AppFileWatcher.cs
@@ -1,5 +1,6 @@
 ﻿using FolderMonitoringService.Config;
 using FolderMonitoringService.Interfaces;
+using Microsoft.Extensions.Logging;
 
 namespace FolderMonitoringService.Services
 {
@@ -7,6 +8,7 @@ namespace FolderMonitoringService.Services
     {
         private readonly List<FileSystemWatcher> FolderWatchers = [];
         private readonly List<FolderMonitorConfig> FolderMonitorConfigs = configuration.GetSection("Folders").Get<List<FolderMonitorConfig>>() ?? [];
+        private const int _maxRetries = 10;
 
         public void Start()
         {
@@ -15,12 +17,12 @@ namespace FolderMonitoringService.Services
         }
         private void StartFileSystemWatcher()
         {
-            // Loop the list to process each of the folder specifications found
+            // Create folder watcher for each configuration
             foreach (FolderMonitorConfig folderConfig in FolderMonitorConfigs)
             {
                 DirectoryInfo dir = new(folderConfig.FolderPath);
 
-                // check if directory is a valid
+                // check if directory is valid
                 if (!dir.Exists)
                 {
                     logger.LogError($"directory not found at {folderConfig.FolderPath}");
@@ -28,7 +30,7 @@ namespace FolderMonitoringService.Services
                 }
 
                 // Checks whether the folder is enabled
-                if (!folderConfig.IsEnabled)
+                if (!folderConfig.Enabled)
                 {
                     logger.LogError($"directory monitoring disabled for {folderConfig.FolderPath}");
                     continue;
@@ -52,33 +54,49 @@ namespace FolderMonitoringService.Services
                 folderWatch.Path = folderConfig.FolderPath;
 
                 // Subscribe to notify filters
-                folderWatch.NotifyFilter = NotifyFilters.DirectoryName
-                             | NotifyFilters.FileName
+                folderWatch.NotifyFilter = NotifyFilters.FileName
                              | NotifyFilters.Size;
 
                 // Associate the event that will be triggered when a new file
                 // is added to the monitored folder, using a lambda expression
-                // TODO creation event not working
+                
                 //folderWatch.Created += (senderObj, fileSysArgs) =>
                 //  OnFileChanged(senderObj, fileSysArgs, folderConfig);
+                
                 folderWatch.Changed += (senderObj, fileSysArgs) =>
                   OnFileChanged(senderObj, fileSysArgs, folderConfig);
+                
                 folderWatch.Renamed += (senderObj, fileSysArgs) =>
                   OnFileChanged(senderObj, fileSysArgs, folderConfig);
+                
                 folderWatch.Error += OnFileWatcherError;
-
 
                 // Begin watching
                 folderWatch.EnableRaisingEvents = true;
 
                 // enable monitoring in sub folders also
-                folderWatch.IncludeSubdirectories = folderConfig.IsIncludeSubFolders;
+                folderWatch.IncludeSubdirectories = folderConfig.IncludeSubFolders;
 
                 // Add the systemWatcher to the list
                 FolderWatchers.Add(folderWatch);
 
                 // Record a log entry into Windows Event Log
                 logger.LogInformation(message: $"Starting to monitor files for allowed extensions {string.Join(", ", folderConfig.AllowedExtensions)} in the folder {folderWatch.Path}");
+            }
+
+            // perform initial folder scan if required
+            foreach (var folderConfig in FolderMonitorConfigs)
+            {
+                if (!folderConfig.InitialScan)
+                {
+                    continue;
+                }
+                // check if directory is a valid
+                if (!Directory.Exists(folderConfig.FolderPath))
+                {
+                    continue;
+                }
+                // TODO process directory
             }
 
         }
@@ -98,8 +116,8 @@ namespace FolderMonitoringService.Services
 
         private void OnFileChanged(object sender, FileSystemEventArgs e, FolderMonitorConfig folderConfig)
         {
+            //logger.LogWarning($"{e.ChangeType}");
             _ = DeleteFileIfViolating(e.FullPath, folderConfig);
-
         }
 
         private bool DeleteFileIfViolating(string filePath, FolderMonitorConfig folderConfig)
@@ -122,7 +140,7 @@ namespace FolderMonitoringService.Services
 
             // delete if file size is more
             var maxAllowedFileSize = folderConfig.MaxFileSizeMb;
-            long fileSizeKb = new FileInfo(filePath).Length / 1024;
+            long fileSizeKb = new FileInfo(filePath).Length / (1024*1024);
             if (fileSizeKb > maxAllowedFileSize)
             {
                 DeleteFile(filePath, $"file size {fileSizeKb}KB is more than {maxAllowedFileSize}KB");
@@ -133,14 +151,59 @@ namespace FolderMonitoringService.Services
 
         private void DeleteFile(string filePath, string deleteMsg)
         {
-            File.Delete(filePath);
-            logger.LogInformation($"{filePath} deleted, {deleteMsg}");
-
+            bool isFileUnLocked = WaitForFile(filePath);
+            if (isFileUnLocked)
+            {
+                File.Delete(filePath);
+                logger.LogWarning($"{filePath} deleted, {deleteMsg}");
+            }
         }
 
         private void OnFileWatcherError(object sender, ErrorEventArgs e)
         {
-            logger.LogInformation($"File error event {e.GetException().Message}");
+            logger.LogError($"File error event {e.GetException().Message}");
+        }
+
+        private bool WaitForFile(string fullPath)
+        {
+            int numTries = 0;
+            while (true)
+            {
+                ++numTries;
+                try
+                {
+                    // Attempt to open the file exclusively.
+                    using (FileStream fs = new FileStream(fullPath,
+                        FileMode.Open, FileAccess.ReadWrite,
+                        FileShare.None, 100))
+                    {
+                        fs.ReadByte();
+                        // If we got this far the file is ready
+                        break;
+                    }
+                }
+                catch (Exception ex)
+                {
+                    logger.LogWarning(
+                       "WaitForFile {0} failed to get an exclusive lock: {1}",
+                        fullPath, ex.ToString());
+
+                    if (numTries > _maxRetries)
+                    {
+                        logger.LogWarning(
+                            "WaitForFile {0} giving up after 10 tries",
+                            fullPath);
+                        return false;
+                    }
+
+                    // Wait for the lock to be released
+                    Thread.Sleep(500);
+                }
+            }
+
+            logger.LogTrace("WaitForFile {0} returning true after {1} tries",
+                fullPath, numTries);
+            return true;
         }
 
     }
